fix syntax error in finishopenidlogin
darcs-hash:20080618123941-84dde-ec6e2fdef8c561714e509dc47e14aa47c58a9c44.gz

+++ actions/finishopenidlogin.php
@@ -264,8 +264,9 @@ class FinishopenidloginAction extends Action {
 			$profile->delete();
 		}
 
+		$result = oid_link_user($user->id, $canonical, $display);
 		
-		if (!oid_link_user($user->id, $canonical, $display) {
+		if (!$result) {
 			# Try to clean up...
 			$user->delete();
 			$profile->delete();
